Add files via upload
This folder contains the simple McIntosh task implemented in PsychoPy. It additionally contains a manual with instructions for installing Psychopy, get it running correctly, and extracting and interpreting data.
This task is a variation of the classic Line Bisection task which was put forward by McIntosh and colleagues. With this task, you can extract an "Endpoint Weightings Bias", which reflects the weighting of the horizontal left and right endpoint positions of the presented lines, as well as an "Endpoint Weightings Sum", which reflects the consistency with which participants responded to different line types.